Simple text file with the audio data properties (flags, psi and loop offset) and minor changes
Improves in the unpacker creating a simple text file with some audio data important properties, that will the EuroSound  will read, includes the following properties:
-Flags
-PSI Sampleheader
-Loop offset.

Apart from this I've added a "marker" in the .yaml files to ensure that EuroSound is reading the correct file.
ftype:1 -> For the Sound bank list.
ftype:2 -> for the effectproperties.yml file.
ftype:3 -> For the audio properties .txt file.